Move scroll pane logic to DetailsController
